# server/api.yaml
openapi: 3.0.0
servers:
  - url: ./
info:
  title: opensvc collector api
  version: 1.0.3
paths:
...
  /filters/{filter_id}:
    post:
      operationId: PostFilter
      description: Modify a filter properties
      parameters:
        - in: path
          name: filter_id
          required: true
          description: Filter identifier (gen_filters.id or f_label)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                f_table:
                  type: string
                f_field:
                  type: string
                f_op:
                  type: string
                f_value:
                  type: string
                f_label:
                  type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteFilter
      description: Delete a filter and its attachments to filtersets
      parameters:
        - in: path
          name: filter_id
          required: true
          description: Filter identifier (gen_filters.id or f_label)
          schema:
            type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    get:
      operationId: GetFilter
      description: Display a filter properties
      parameters:
        - in: path
          name: filter_id
          required: true
          description: Filter identifier (gen_filters.id)
          schema:
            type: string
        - $ref: '#/components/parameters/inQueryProps'
        - $ref: '#/components/parameters/inQueryLimit'
        - $ref: '#/components/parameters/inQueryOffset'
        - $ref: '#/components/parameters/inQueryMeta'
        - $ref: '#/components/parameters/inQueryStats'
        - $ref: '#/components/parameters/inQueryOrderby'
        - $ref: '#/components/parameters/inQueryGroupby'
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
...
  /filtersets/{filterset_id}:
    post:
      operationId: PostFilterset
      description: Modify a filterset properties
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                fset_name:
                  type: string
                fset_stats:
                  type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteFilterset
      description: Delete a filterset and its attachments and references
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    get:
      operationId: GetFilterset
      description: Display a filterset properties
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
        - $ref: '#/components/parameters/inQueryProps'
        - $ref: '#/components/parameters/inQueryLimit'
        - $ref: '#/components/parameters/inQueryOffset'
        - $ref: '#/components/parameters/inQueryMeta'
        - $ref: '#/components/parameters/inQueryStats'
        - $ref: '#/components/parameters/inQueryOrderby'
        - $ref: '#/components/parameters/inQueryGroupby'
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
...
  /filtersets/{filterset_id}/filters/{f_id}:
    post:
      operationId: PostFiltersetFilter
      description: Attach a filter to a filterset
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
        - in: path
          name: f_id
          required: true
          description: Filter identifier (gen_filters.id or f_label)
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                f_order:
                  type: integer
                f_log_op:
                  type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteFiltersetFilter
      description: Detach a filter from a filterset
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
        - in: path
          name: f_id
          required: true
          description: Filter identifier (gen_filters.id or f_label)
          schema:
            type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
...
components:
  responses:
...
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    ListMeta:
      type: object
      required:
      properties:
        count:
          type: integer
        available_props:
          type: array
          items:
            type: string
        included_props:
          type: array
          items:
            type: string
        limit:
          type: integer
        offset:
          type: integer
        distinct:
          type: object
          additionalProperties:
            type: integer
        total:
          type: integer
    ListResponse:
      type: object
      required:
        - data
      properties:
        data:
          oneOf:
            - type: array
              items:
                type: object
            - type: object
              additionalProperties:
                type: object
                additionalProperties:
                  type: integer
        meta:
          $ref: '#/components/schemas/ListMeta'
    Problem:
      type: object
      properties:
        text:
          description: |
            A human-readable explanation specific to this occurrence of the
            problem.
          type: string
      required:
        - text
...
  parameters:
...
    inQueryGroupby:
      in: query
      name: groupby
      required: false
      description: Comma-separated list of properties to group the result by (e.g. groupby=app,svcname).
      schema:
        type: string
    inQueryLimit:
      in: query
      name: limit
      required: false
      description: The maximum number of entries to return. 0 means no limit.
      schema:
        type: integer
        minimum: 0
        default: 50
    inQueryMeta:
      in: query
      name: meta
      required: false
      description: Include metadata in the response. Enabled by default. Use false or 0 to omit the meta field.
      schema:
        type: string
    inQueryOffset:
      in: query
      name: offset
      required: false
      description: Skip the first entries of the data cursor.
      schema:
        type: integer
        minimum: 0
        default: 0
    inQueryOrderby:
      in: query
      name: orderby
      required: false
      description: Comma-separated list of properties to sort by. Prefix a property with - for descending order (e.g. orderby=nodename,-app).
      schema:
        type: string
    inQueryProps:
      in: query
      name: props
      required: false
      description: A list of properties to include in each data dictionnary.
      schema:
        type: string
    inQueryStats:
      in: query
      name: stats
      required: false
      description: Controls the inclusion in the returned dictionnary of a "stats" key, containing the selected properties distinct values counts.
      schema:
        type: string
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT